# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactSuccess {
  ok: boolean;
  message: string;
}

export interface ContactError {
  ok: boolean;
  error: string;
}

export type PartnerInquiryInputInquiryType = typeof PartnerInquiryInputInquiryType[keyof typeof PartnerInquiryInputInquiryType];


export const PartnerInquiryInputInquiryType = {
  discovery: 'discovery',
  pilot: 'pilot',
  institutional_partnership: 'institutional_partnership',
} as const;

export type PartnerInquiryInputHasClinicalOwner = typeof PartnerInquiryInputHasClinicalOwner[keyof typeof PartnerInquiryInputHasClinicalOwner];


export const PartnerInquiryInputHasClinicalOwner = {
  yes: 'yes',
  no: 'no',
  in_discussion: 'in_discussion',
} as const;

export type PartnerInquiryInputHasIntermediaries = typeof PartnerInquiryInputHasIntermediaries[keyof typeof PartnerInquiryInputHasIntermediaries];


export const PartnerInquiryInputHasIntermediaries = {
  yes: 'yes',
  no: 'no',
  in_discussion: 'in_discussion',
} as const;

export type PartnerInquiryInputHasReferralPathway = typeof PartnerInquiryInputHasReferralPathway[keyof typeof PartnerInquiryInputHasReferralPathway];


export const PartnerInquiryInputHasReferralPathway = {
  yes: 'yes',
  no: 'no',
  unclear: 'unclear',
} as const;

export type PartnerInquiryInputCaseVolumeBand = typeof PartnerInquiryInputCaseVolumeBand[keyof typeof PartnerInquiryInputCaseVolumeBand];


export const PartnerInquiryInputCaseVolumeBand = {
  fewer_than_10: 'fewer_than_10',
  '10_to_49': '10_to_49',
  '50_to_199': '50_to_199',
  '200_plus': '200_plus',
  unknown: 'unknown',
} as const;

/**
 * @nullable
 */
export type PartnerInquiryInputFundingStatus = typeof PartnerInquiryInputFundingStatus[keyof typeof PartnerInquiryInputFundingStatus] | null;


export const PartnerInquiryInputFundingStatus = {
  funded: 'funded',
  seeking_sponsor: 'seeking_sponsor',
  internal_budget_possible: 'internal_budget_possible',
  unknown: 'unknown',
} as const;

export interface PartnerInquiryInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  fullName: string;
  /**
     * @minLength 5
     * @maxLength 254
     */
  workEmail: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  organization: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  role: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  organizationType: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  country: string;
  /**
     * @maxLength 500
     * @nullable
     */
  organizationWebsite?: string | null;
  inquiryType: PartnerInquiryInputInquiryType;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  workflowChallenge: string;
  hasClinicalOwner: PartnerInquiryInputHasClinicalOwner;
  hasIntermediaries: PartnerInquiryInputHasIntermediaries;
  hasReferralPathway: PartnerInquiryInputHasReferralPathway;
  caseVolumeBand: PartnerInquiryInputCaseVolumeBand;
  /**
     * @minLength 1
     * @maxLength 500
     */
  preferredNextStep: string;
  /**
     * @maxLength 500
     * @nullable
     */
  linkedinProfile?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  publicProgramPage?: string | null;
  /**
     * @maxLength 200
     * @nullable
     */
  targetPilotTiming?: string | null;
  /** @nullable */
  fundingStatus?: PartnerInquiryInputFundingStatus;
  /**
     * @maxLength 500
     * @nullable
     */
  howHeardAbout?: string | null;
  consentGiven: boolean;
  noSensitiveDataConfirmed: boolean;
  submittedAt: string;
  /**
     * Must be empty — bot trap field
     * @nullable
     */
  honeypot?: string | null;
}

export type AdvisorInterestInputInterestType = typeof AdvisorInterestInputInterestType[keyof typeof AdvisorInterestInputInterestType];


export const AdvisorInterestInputInterestType = {
  pro_bono_advisory: 'pro_bono_advisory',
  paid_advisory: 'paid_advisory',
  research_collaboration: 'research_collaboration',
  institutional_introduction: 'institutional_introduction',
} as const;

export interface AdvisorInterestInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  fullName: string;
  /**
     * @minLength 5
     * @maxLength 254
     */
  professionalEmail: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  currentRole: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  organization: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  expertiseArea: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  country: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  experienceSummary: string;
  /**
     * @maxLength 500
     * @nullable
     */
  professionalProfile?: string | null;
  /**
     * @minLength 1
     * @maxLength 200
     */
  availabilityBand: string;
  interestType: AdvisorInterestInputInterestType;
  /** @maxLength 1000 */
  conflictDisclosure: string;
  consentGiven: boolean;
  noSensitiveDataConfirmed: boolean;
  submittedAt: string;
  /**
     * Must be empty — bot trap field
     * @nullable
     */
  honeypot?: string | null;
}

export type GeneralContactInputInquiryType = typeof GeneralContactInputInquiryType[keyof typeof GeneralContactInputInquiryType];


export const GeneralContactInputInquiryType = {
  host_partner: 'host_partner',
  pilot_sponsor: 'pilot_sponsor',
  accelerator: 'accelerator',
  clinical_advisor: 'clinical_advisor',
  security_advisor: 'security_advisor',
  legal_advisor: 'legal_advisor',
  evaluation_partner: 'evaluation_partner',
  technology_partner: 'technology_partner',
  media: 'media',
  general: 'general',
} as const;

export interface GeneralContactInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /**
     * @minLength 5
     * @maxLength 254
     */
  workEmail: string;
  /**
     * @maxLength 300
     * @nullable
     */
  organization?: string | null;
  inquiryType: GeneralContactInputInquiryType;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  consentGiven: boolean;
  noSensitiveDataConfirmed: boolean;
  submittedAt: string;
  /**
     * Must be empty — bot trap field
     * @nullable
     */
  honeypot?: string | null;
}