# lib/api-zod/src/generated/types/advisorInterestInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdvisorInterestInputInterestType } from './advisorInterestInputInterestType';

export interface AdvisorInterestInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  fullName: string;
  /**
     * @minLength 5
     * @maxLength 254
     */
  professionalEmail: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  currentRole: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  organization: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  expertiseArea: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  country: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  experienceSummary: string;
  /**
     * @maxLength 500
     * @nullable
     */
  professionalProfile?: string | null;
  /**
     * @minLength 1
     * @maxLength 200
     */
  availabilityBand: string;
  interestType: AdvisorInterestInputInterestType;
  /** @maxLength 1000 */
  conflictDisclosure: string;
  consentGiven: boolean;
  noSensitiveDataConfirmed: boolean;
  submittedAt: string;
  /**
     * Must be empty — bot trap field
     * @nullable
     */
  honeypot?: string | null;
}