# lib/api-zod/src/generated/types/generalContactInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneralContactInputInquiryType } from './generalContactInputInquiryType';

export interface GeneralContactInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /**
     * @minLength 5
     * @maxLength 254
     */
  workEmail: string;
  /**
     * @maxLength 300
     * @nullable
     */
  organization?: string | null;
  inquiryType: GeneralContactInputInquiryType;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  consentGiven: boolean;
  noSensitiveDataConfirmed: boolean;
  submittedAt: string;
  /**
     * Must be empty — bot trap field
     * @nullable
     */
  honeypot?: string | null;
}