# lib/api-zod/src/generated/types/partnerInquiryInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerInquiryInputCaseVolumeBand } from './partnerInquiryInputCaseVolumeBand';
import type { PartnerInquiryInputFundingStatus } from './partnerInquiryInputFundingStatus';
import type { PartnerInquiryInputHasClinicalOwner } from './partnerInquiryInputHasClinicalOwner';
import type { PartnerInquiryInputHasIntermediaries } from './partnerInquiryInputHasIntermediaries';
import type { PartnerInquiryInputHasReferralPathway } from './partnerInquiryInputHasReferralPathway';
import type { PartnerInquiryInputInquiryType } from './partnerInquiryInputInquiryType';

export interface PartnerInquiryInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  fullName: string;
  /**
     * @minLength 5
     * @maxLength 254
     */
  workEmail: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  organization: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  role: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  organizationType: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  country: string;
  /**
     * @maxLength 500
     * @nullable
     */
  organizationWebsite?: string | null;
  inquiryType: PartnerInquiryInputInquiryType;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  workflowChallenge: string;
  hasClinicalOwner: PartnerInquiryInputHasClinicalOwner;
  hasIntermediaries: PartnerInquiryInputHasIntermediaries;
  hasReferralPathway: PartnerInquiryInputHasReferralPathway;
  caseVolumeBand: PartnerInquiryInputCaseVolumeBand;
  /**
     * @minLength 1
     * @maxLength 500
     */
  preferredNextStep: string;
  /**
     * @maxLength 500
     * @nullable
     */
  linkedinProfile?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  publicProgramPage?: string | null;
  /**
     * @maxLength 200
     * @nullable
     */
  targetPilotTiming?: string | null;
  /** @nullable */
  fundingStatus?: PartnerInquiryInputFundingStatus;
  /**
     * @maxLength 500
     * @nullable
     */
  howHeardAbout?: string | null;
  consentGiven: boolean;
  noSensitiveDataConfirmed: boolean;
  submittedAt: string;
  /**
     * Must be empty — bot trap field
     * @nullable
     */
  honeypot?: string | null;
}